refactor(component): refactor text components(type, props)
- StrongText의 경우 props로 size를 추가했다.
- 타입에 대해 공통 모듈에 넣어주었다.
