+++ packages/ui/text/Strong.tsx
@@ -1,22 +1,20 @@
-import { PropsWithChildren } from 'react';
-
 import styled from '@emotion/styled';
 
 import { globalTheme } from '@ui/styles';
 
-export type StrongColorType = 'primary' | 'sub' | 'black' | 'white' | 'inherit';
+import { StrongTextPropsInterface, StyledStrongInterface } from './types';
 
-interface StrongTextPropsInterface extends PropsWithChildren {
-  color?: StrongColorType & string;
-}
-
-const StyledStrong = styled.strong<{
-  color: StrongColorType | string;
-}>`
+const StyledStrong = styled.strong<StyledStrongInterface>`
+  font-size: ${(props) => props.size};
   color: ${(props) => props.color};
 `;
 
-export const StrongText = ({ color = 'black', children, ...props }: StrongTextPropsInterface) => {
+export const StrongText = ({
+  size = globalTheme.fontSize.md,
+  color = 'black',
+  children,
+  ...props
+}: StrongTextPropsInterface) => {
   const colors = {
     primary: globalTheme.color.primary[500],
     sub: globalTheme.color.sub[500],
@@ -26,7 +24,7 @@ export const StrongText = ({ color = 'black', children, ...props }: StrongTextPr
   };
 
   return (
-    <StyledStrong {...props} color={colors[color] ?? color}>
+    <StyledStrong {...props} size={size} color={colors[color] ?? color}>
       {children}
     </StyledStrong>
   );
